backend/requirements-desktop-macos.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
yfinance==0.2.66
pandas==2.2.0
pandas-market-calendars>=4.5.0
numpy==1.26.3
requests==2.31.0
python-multipart==0.0.6
finvizfinance==1.2.0
openpyxl==3.1.2
scipy==1.13.1
ruptures>=1.1.9
statsmodels>=0.14.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
tzdata>=2024.1
pyinstaller>=6.11.0
pywebview>=5.3.2
pytest>=8.0.0
pytest-asyncio>=0.23.0
